+++ docs/scripts/views/guide/upgrade.vue
@@ -1,0 +1,14 @@
+<template>
+  <pure-page name="upgrade"></pure-page>
+</template>
+
+<script>
+export default {
+  metaInfo: {
+    titleTemplate: '%s - Upgrade Guide'
+  },
+  data() {
+    return {};
+  }
+};
+</script>